feat: add operation system module
